move peer certificate verification to new file
