+++ examples/client/websocket/without_tls/main.go
@@ -48,7 +48,7 @@ func main() {
 	fmt.Printf("Failed to connect with invalid path %s,error : %s\n", invalidPathBrokerAddress, err.Error())
 }
 
-func onMessage(c mqtt.Client, m mqtt.Message, done chan struct{}) {
+func onMessage(_ mqtt.Client, m mqtt.Message, done chan struct{}) {
 	fmt.Printf("Subscription Message Received, Topic : %s, Payload %s\n", m.Topic(), string(m.Payload()))
 	done <- struct{}{}
 }
